Finished 4-bit design
Program now works for 4-bit resolution on the Uno. It can control 12-13 tiles before the lights start to dim and the program speed slows noticeably. At 16 tiles (1/2 the board) it does not work at all. Next going to experiment with a 3-bit design, fiddle with slowing the communication, and getting it to work on the Mega.